trojkat sie obraca

+++ STR.h
@@ -11,7 +11,7 @@ struct CircleData {
 };
 
 struct TriangleData {
-	int x0, y0, x1, y1, x2, y2;
+	float x0, y0, x1, y1, x2, y2;
     ALLEGRO_COLOR color;
 };
